Fixing total entries error for group by on field other than id with count or sum in select
With https://github.com/drewolson/scrivener_ecto/pull/39, I partially
fixed the issue with grouping by field other than id, but then I
realized that it would still not work if there was a count or sum in the
select.

The is a more complete solution. If there is a group by clause, it
uses the first column in the group by clause for the total entries
select.

Fixes #35
Fixes #39